JWT v0.1.8 jwt

JWT Library for Erlang.

Written by Peter Hizalev at Kato (http://kato.im)

Rewritten by Yuri Artemev (http://artemff.com)

Functions

algorithm_to_crypto(_)

algorithm_to_crypto(binary()) :: {atom(), atom()} | :undefined

append_claim(ClaimsSet, Key, Val)

append_claim(claimsSet :: map() | list(), binary(), any()) :: map() | list()

decode(Token, Key)

decode(
  token :: binary(),
  key :: binary() | :public_key.public_key() | :public_key.private_key()
) :: {:ok, claims :: map()} | {:error, any()}

Decodes a token, checks the signature and returns the content of the token

Token is a JWT itself

Key is a secret phrase or public/private key depend on encryption algorithm

decode(Token, DefaultKey, IssuerKeyMapping)

decode(
  token :: binary(),
  defaultKey :: binary() | :public_key.public_key() | :public_key.private_key(),
  issuerKeyMapping :: map()
) :: {:ok, claims :: map()} | {:error, any()}

Decode with an issuer key mapping

Receives the issuer key mapping as the last parameter

decode_jwt(Context)

decode_jwt(context()) :: {true, context()} | {false, {:error, :invalid_token}}

encode(Alg, ClaimsSet, Key)

encode(
  alg :: binary(),
  claimsSet :: map() | list(),
  key :: binary() | :public_key.private_key()
) :: {:ok, token :: binary()} | {:error, any()}

Creates a token from given data and signs it with a given secret

Parameters are*

Alg is a binary one of

[HS256, HS384, HS512, RS256, RS384, RS512, ES256, ES384, ES512, PS256, PS384, PS512]

But only [HS256, HS384, HS512, RS256] are supported

ClaimsSet the payload of the token. Can be both map and proplist

Key binary in case of hmac encryption and private key if rsa

encode(Alg, ClaimsSet, Expiration, Key)

encode(
  alg :: binary(),
  claimsSet :: map() | list(),
  expiration :: expiration(),
  key :: binary() | :public_key.private_key()
) :: {:ok, token :: binary()} | {:error, any()}

Creates a token from given data and signs it with a given secret

and also adds exp claim to payload

Expiration can be one of the tuples:{hourly, SecondsAfterBeginningOfCurrentHour},{daily, SecondsAfterBeginningOfCurrentDay} or can be just an integer representing the amount of seconds the token will live

epoch()

epoch() :: non_neg_integer()

expiration_to_epoch(Expiration)

expiration_to_epoch(expiration :: expiration()) :: neg_integer()

expiration_to_epoch(Expiration, Ts)

jsx_decode_safe(Bin)

jsx_decode_safe(binary()) :: map() | :invalid

jwt_add_exp(ClaimsSet, Expiration)

jwt_add_exp(claimsSet :: map() | list(), expiration :: expiration()) ::
  map() | list()

jwt_check_sig(Alg, Payload, Signature, Key)

jwt_check_sig(
  {atom(), atom()},
  payload :: binary(),
  signature :: binary(),
  key :: binary() | :public_key.public_key() | :public_key.private_key()
) :: boolean()

jwt_check_sig(Alg, Header, Claims, Signature, Key)

jwt_check_sig(
  alg :: binary(),
  header :: binary(),
  claims :: binary(),
  signature :: binary(),
  key :: binary() | :public_key.public_key() | :public_key.private_key()
) :: boolean()

jwt_header(Alg)

jwt_header(alg :: binary()) :: list()

jwt_is_expired(ClaimsJSON)

jwt_is_expired(map()) :: boolean()

jwt_sign(Alg, Payload, Key)

jwt_sign(
  alg :: binary(),
  payload :: binary(),
  key :: binary() | :public_key.private_key()
) :: binary() | :undefined

jwt_sign_with_crypto(_, Payload, Key)

pem_to_key(Pem)

result(Error)

run(Acc, Funs)

split_token(Context)

split_token(context :: context()) ::
  {true, context()} | {false, {:error, :invalid_token}}